// packages/core/config.d.ts
export declare const config: {
    AGENT_MAX_ITERATIONS: number;
    CODE_EXECUTION_TIMEOUT_MS: number;
    CONTAINER_MEMORY_LIMIT: string;
    HISTORY_MAX_LENGTH: number;
    HOST_PROJECT_PATH: string;
    LLM_API_KEY?: string | undefined;
    LLM_MODEL_NAME: string;
    MCP_API_KEY?: string | undefined;
    MCP_WEBHOOK_URL?: string | undefined;
    NODE_ENV: string;
    PORT: number;
    PYTHON_SANDBOX_IMAGE: string;
    QUALITY_GATE_API_KEY?: string | undefined;
    QUALITY_GATE_URL?: string | undefined;
    REDIS_HOST: string;
    REDIS_PORT: number;
    WEBHOOK_SECRET?: string | undefined;
    WORKER_CONCURRENCY: number;
};
//# sourceMappingURL=config.d.ts.map